pyproject.toml
[project]
name = "voice-orchestrator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepgram-sdk>=4.8.1",
    "dotenv>=0.9.9",
    "elevenlabs>=2.16.0",
    "evdev-binary>=1.9.2",
    "numpy>=2.3.3",
    "openai>=1.108.1",
    "pydantic-settings>=2.10.1",
    "pynput>=1.8.1",
    "requests>=2.32.5",
    "sounddevice>=0.5.2",
    "websockets>=15.0.1",
]
